🚚RENAME: preprocessing folder to `preprocessing`
**MODIFY:**
- 🧰 `Makefile`: change deprency place
- 🚚`src/main.c`: change include path
- 🚚`src/segmentation/auto_rotation.h`: change include path

**RENAME:**
- 🚚`src/load_Image/*` -> `src/preprocessing/*` : match module name
