+++ src/preprocessing/change_color.c
