+++ src/preprocessing/change_color.h
